feat, fix(redis cloud, integration with members,subscription): create new funtionally, modify entity member and integration of redis cloud    - change all atributes of the entity member with Integer to Long      - create a new endpoint where the member can see the state of thesubscription.

+++ src/main/java/com/GymManager/Backend/domain/repository/GymMemberPersistencePort.java
@@ -12,15 +12,15 @@ import java.util.Optional;
 
 public interface GymMemberPersistencePort {
     GymMembers save(GymMembers gymMember);
-    Optional<GymMembers> findById(Integer id);
-    Optional<GymMemberFullData> getFullData(Integer id);
-    Optional<GymMembers> findByIdentificationNumber(Integer id);
+    Optional<GymMembers> findById(Long id);
+    Optional<GymMemberFullData> getFullData(Long id);
+    Optional<GymMembers> findByIdentificationNumber(Long id);
     List<GymMembers> findAll();
     List<GymMembers> findAllByParam(String param);
-    void deleteById(Integer id);
+    void deleteById(Long id);
     GymMembers update(GymMembers gymMember);
-    Boolean existById(Integer id);
-    Boolean existByIdentification(Integer id);
+    Boolean existById(Long id);
+    Boolean existByIdentification(Long id);
     List<AverageGenderDistributionDto> findTotalMembersByGender();
     List<AllDataAboutUser> findAllLastRegisteredMembers();
 }
