+++ src/main/java/com/GymManager/Backend/persistence/JpaRepositoriImpl/GymMemberAdapter.java
@@ -45,12 +45,12 @@ public class GymMemberAdapter implements GymMemberPersistencePort {
     }
 
     @Override
-    public Optional<GymMembers> findById(Integer id) {
+    public Optional<GymMembers> findById(Long id) {
         return gymMemberCrudRepo.findById(id);
     }
 
     @Override
-    public Optional<GymMemberFullData> getFullData(Integer id) {
+    public Optional<GymMemberFullData> getFullData(Long id) {
         AllDataAboutUser allData = this.gymMemberCrudRepo.allDataOfUser(id);
         if (allData == null) {
             return Optional.empty();
@@ -75,7 +75,7 @@ public class GymMemberAdapter implements GymMemberPersistencePort {
     }
 
     @Override
-    public Optional<GymMembers> findByIdentificationNumber(Integer id) {
+    public Optional<GymMembers> findByIdentificationNumber(Long id) {
         return this.gymMemberCrudRepo.findByIdentificationNumber(id);
     }
 
@@ -91,7 +91,7 @@ public class GymMemberAdapter implements GymMemberPersistencePort {
 
     @Override
     @Transactional
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         this.subscriptionPersistencePort.findByMember_IdMember(id).ifPresent(this.subscriptionPersistencePort::delete);
         gymMemberCrudRepo.deleteById(id);
     }
@@ -102,12 +102,12 @@ public class GymMemberAdapter implements GymMemberPersistencePort {
     }
 
     @Override
-    public Boolean existById(Integer id) {
+    public Boolean existById(Long id) {
         return this.gymMemberCrudRepo.existsById(id);
     }
 
     @Override
-    public Boolean existByIdentification(Integer id) {
+    public Boolean existByIdentification(Long id) {
         return this.gymMemberCrudRepo.existsByIdentificationNumber(id);
     }
